+++ plugins/providers/docker/action.rb
@@ -215,7 +215,7 @@ module VagrantPlugins
       def self.action_start
         Vagrant::Action::Builder.new.tap do |b|
           b.use Call, IsState, :running do |env, b2|
-            # If the container is running and we're doing a run, we're done
+            # If the container is running and we're not doing a run, we're done
             next if env[:result] && env[:machine_action] != :run_command
 
             if env[:machine_action] != :run_command
@@ -231,7 +231,10 @@ module VagrantPlugins
             end
 
             b2.use Call, IsState, :not_created do |env2, b3|
-              if !env2[:result]
+              if env2[:result]
+                # First time making this thing, set to the "preparing" state
+                b3.use InitState
+              else
                 b3.use EnvSet, host_machine_sync_folders: false
               end
             end
@@ -247,7 +250,7 @@ module VagrantPlugins
             if env[:machine_action] != :run_command
               # If the container is NOT created yet, then do some setup steps
               # necessary for creating it.
-              b2.use Call, IsState, :not_created do |env2, b3|
+              b2.use Call, IsState, :preparing do |env2, b3|
                 if env2[:result]
                   b3.use EnvSet, port_collision_repair: true
                   b3.use HostMachinePortWarning
@@ -271,6 +274,7 @@ module VagrantPlugins
                 end
               end
             else
+              # We're in a run command, so we do things a bit differently.
               b2.use SyncedFolders
               b2.use Create
             end
@@ -294,6 +298,7 @@ module VagrantPlugins
       autoload :HostMachineRequired, action_root.join("host_machine_required")
       autoload :HostMachineSyncFolders, action_root.join("host_machine_sync_folders")
       autoload :HostMachineSyncFoldersDisable, action_root.join("host_machine_sync_folders_disable")
+      autoload :InitState, action_root.join("init_state")
       autoload :IsBuild, action_root.join("is_build")
       autoload :IsHostMachineCreated, action_root.join("is_host_machine_created")
       autoload :Login, action_root.join("login")